Remove platform specific structure definitions
Instead of having less/more fields inside a structure depending on the
platform (nonlinux/linux), it would be better to have the same standard
definition in all platforms, and certain fields of the structure to be
populated or left unpopulated depending on the platform.

This will make things easier for other projects like `mc` to use these
types without having to worry about them having platform-dependent
structure.
